refactor: Move filtersToQuery to common-utils
