+++ packages/app/src/searchFilters.tsx
@@ -1,64 +1,15 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import produce from 'immer';
+import {
+  type FilterState,
+  filtersToQuery,
+} from '@hyperdx/common-utils/dist/filters';
 import type { Filter } from '@hyperdx/common-utils/dist/types';
 
 import { usePinnedFiltersApi, useUpdatePinnedFilters } from './pinnedFilters';
 import { useLocalStorage } from './utils';
 
 export const IS_ROOT_SPAN_COLUMN_NAME = 'isRootSpan';
-
-export type FilterState = {
-  [key: string]: {
-    included: Set<string | boolean>;
-    excluded: Set<string | boolean>;
-    range?: { min: number; max: number }; // For BETWEEN conditions
-  };
-};
-
-export const filtersToQuery = (
-  filters: FilterState,
-  { stringifyKeys = false }: { stringifyKeys?: boolean } = {},
-): Filter[] => {
-  return Object.entries(filters)
-    .filter(
-      ([_, values]) =>
-        values.included.size > 0 ||
-        values.excluded.size > 0 ||
-        values.range != null,
-    )
-    .flatMap(([key, values]) => {
-      const conditions = [];
-      const actualKey = stringifyKeys ? `toString(${key})` : key;
-
-      if (values.included.size > 0) {
-        conditions.push({
-          type: 'sql' as const,
-          condition: `${actualKey} IN (${Array.from(values.included)
-            .map(v =>
-              typeof v === 'string' ? `'${v.replace(/'/g, "''")}'` : v,
-            )
-            .join(', ')})`,
-        });
-      }
-      if (values.excluded.size > 0) {
-        conditions.push({
-          type: 'sql' as const,
-          condition: `${actualKey} NOT IN (${Array.from(values.excluded)
-            .map(v =>
-              typeof v === 'string' ? `'${v.replace(/'/g, "''")}'` : v,
-            )
-            .join(', ')})`,
-        });
-      }
-      if (values.range != null) {
-        conditions.push({
-          type: 'sql' as const,
-          condition: `${actualKey} BETWEEN ${values.range.min} AND ${values.range.max}`,
-        });
-      }
-      return conditions;
-    });
-};
 
 export const areFiltersEqual = (a: FilterState, b: FilterState) => {
   const aKeys = Object.keys(a);
