+++ packages/common-utils/src/filters.ts
@@ -1,0 +1,54 @@
+import { Filter } from '@/types';
+
+export type FilterState = {
+  [key: string]: {
+    included: Set<string | boolean>;
+    excluded: Set<string | boolean>;
+    range?: { min: number; max: number }; // For BETWEEN conditions
+  };
+};
+
+export const filtersToQuery = (
+  filters: FilterState,
+  { stringifyKeys = false }: { stringifyKeys?: boolean } = {},
+): Filter[] => {
+  return Object.entries(filters)
+    .filter(
+      ([_, values]) =>
+        values.included.size > 0 ||
+        values.excluded.size > 0 ||
+        values.range != null,
+    )
+    .flatMap(([key, values]) => {
+      const conditions: Filter[] = [];
+      const actualKey = stringifyKeys ? `toString(${key})` : key;
+
+      if (values.included.size > 0) {
+        conditions.push({
+          type: 'sql' as const,
+          condition: `${actualKey} IN (${Array.from(values.included)
+            .map(v =>
+              typeof v === 'string' ? `'${v.replace(/'/g, "''")}'` : v,
+            )
+            .join(', ')})`,
+        });
+      }
+      if (values.excluded.size > 0) {
+        conditions.push({
+          type: 'sql' as const,
+          condition: `${actualKey} NOT IN (${Array.from(values.excluded)
+            .map(v =>
+              typeof v === 'string' ? `'${v.replace(/'/g, "''")}'` : v,
+            )
+            .join(', ')})`,
+        });
+      }
+      if (values.range != null) {
+        conditions.push({
+          type: 'sql' as const,
+          condition: `${actualKey} BETWEEN ${values.range.min} AND ${values.range.max}`,
+        });
+      }
+      return conditions;
+    });
+};
